simplify java example to demonstrate a simple case
