+++ examples/java/.gitignore
